make the TTsize a long
git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@3983 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/application/kdm/src/structures/migrationjob.h
@@ -208,7 +208,7 @@ public:
   /// SDPA VM parameters
   int ReqVMMemSize;
 
-  int globTTbufsizelocal;
+  long globTTbufsizelocal;
 
   int locbufsize;	  // Size of the local buffers;
